requirements.txt
Flask==2.3.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.10.1
matplotlib==3.7.2
seaborn==0.12.2
google-generativeai==0.3.2
